AI/requirements.txt
tensorflow
numpy
pandas
seaborn
matplotlib
cloudinary
google-api-core
ElevenLabs
python-dotenv